+++ tests/tripleo/intr/utils/test_urls.py
@@ -15,7 +15,7 @@
 """
 from unittest import TestCase
 
-from tripleo.utils.types import URL
+from tripleo.insights.types import URL
 from tripleo.utils.urls import is_github, is_git
