+++ tripleo/insights/types.py
